Rewrite for Webnative 0.35
