+++ .github/workflows/develop.yml
@@ -4,33 +4,23 @@ name: Deploy to staging
 on:
   push:
     branches: [ develop ]
-    
+
   workflow_dispatch:
-    
-    
+
+
 jobs:
   build-and-publish:
     runs-on: ubuntu-latest
     steps:
-      
+
       - name: Check out repository
         uses: actions/checkout@v2
-      
+
       - name: Install nix
         uses: cachix/install-nix-action@v12
         with:
           nix_path: nixpkgs=channel:nixos-unstable
-          
-      - name: Cache pnpm modules
-        uses: actions/cache@v2
-        env:
-          cache-name: cache-pnpm-modules
-        with:
-          path: ~/.pnpm-store
-          key: ${{ runner.os }}-build-${{ env.cache-name }}-${{ hashFiles('**/package.json') }}
-          restore-keys: |
-            ${{ runner.os }}-build-${{ env.cache-name }}-
-      
+
       - name: Cache elm home
         uses: actions/cache@v2
         env:
@@ -41,10 +31,10 @@ jobs:
 
       - name: Install Dependencies
         run: nix-shell --run "just install-deps"
-        
+
       - name: Build
         run: nix-shell --run "just staging-build"
-        
+
       - name: Publish to Fission
         uses: fission-suite/publish-action@v1
         with:
