+++ package.json
@@ -6,12 +6,13 @@
     "ipfs-message-port-protocol": "^0.11.3",
     "ipfs-message-port-server": "^0.11.3",
     "localforage": "^1.9.0",
-    "webnative": "^0.34.2"
+    "uint8arrays": "^3.0.0",
+    "webnative": "file:../webnative"
   },
   "devDependencies": {
     "autoprefixer": "^10.4.8",
     "elm-tailwind-css": "^1.1.1",
-    "esbuild": "^0.12.26",
+    "esbuild": "^0.15.15",
     "mustache": "^4.0.1",
     "postcss": "^8.4.16",
     "postcss-import": "^12.0.1",
